Moved device auth calls into top level component, close expanded details on refresh Changelog: none
